requirements/dev.txt
-r common.txt

black==23.1.0
coverage==7.2.2
debugpy==1.6.6
django-extensions==3.2.1
django-debug-toolbar==3.8.1
djhtml==3.0.6
factory-boy==3.2.1
flake8==6.0.0
isort==5.12.0
pylint==2.17.1
pylint-django==2.5.3
python-dotenv==1.0.0
